requirements.txt
requests
beautifulsoup4
httpx
fake-useragent
python-dotenv
pandas
gspread
oauth2client
